# packages/core/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Inspect API",
    "version": "0.1.0",
    "description": "AI-Powered Browser Testing Platform API",
    "license": { "name": "MIT" }
  },
  "servers": [{ "url": "http://localhost:4100", "description": "Local development" }],
  "security": [{ "BearerAuth": [] }],
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Health check",
        "tags": ["System"],
        "security": [],
        "responses": {
          "200": {
            "description": "Healthy",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/HealthResponse" } }
            }
          },
          "503": { "description": "Degraded/Unhealthy" }
        }
      }
    },
    "/api/version": {
      "get": {
        "summary": "Version info",
        "tags": ["System"],
        "security": [],
        "responses": { "200": { "description": "Version details" } }
      }
    },
    "/api/models": {
      "get": {
        "summary": "Supported LLM models",
        "tags": ["System"],
        "security": [],
        "parameters": [{ "name": "provider", "in": "query", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "List of models" } }
      }
    },
    "/api/tasks": {
      "post": {
        "summary": "Create and run a task",
        "tags": ["Tasks"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateTask" } }
          }
        },
        "responses": {
          "201": {
            "description": "Task created",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/TaskResponse" } }
            }
          },
          "400": { "description": "Validation error" }
        }
      }
    },
    "/api/tasks/{id}": {
      "get": {
        "summary": "Get task status",
        "tags": ["Tasks"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Task details" },
          "404": { "description": "Task not found" }
        }
      }
    },
    "/api/tasks/{id}/cancel": {
      "post": {
        "summary": "Cancel a running task",
        "tags": ["Tasks"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Task cancelled" },
          "400": { "description": "Cannot cancel task in current state" },
          "404": { "description": "Task not found" }
        }
      }
    },
    "/api/tasks/{id}/artifacts": {
      "get": {
        "summary": "Get task artifacts",
        "tags": ["Tasks"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Task artifacts" },
          "404": { "description": "Task not found" }
        }
      }
    },
    "/api/workflows": {
      "post": {
        "summary": "Create a workflow",
        "tags": ["Workflows"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateWorkflow" } }
          }
        },
        "responses": {
          "201": { "description": "Workflow created" },
          "400": { "description": "Validation error" }
        }
      },
      "get": {
        "summary": "List workflows",
        "tags": ["Workflows"],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": { "type": "string", "enum": ["draft", "active", "paused", "archived"] }
          }
        ],
        "responses": { "200": { "description": "List of workflows" } }
      }
    },
    "/api/workflows/{id}": {
      "get": {
        "summary": "Get a workflow",
        "tags": ["Workflows"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Workflow details" },
          "404": { "description": "Not found" }
        }
      },
      "put": {
        "summary": "Update a workflow",
        "tags": ["Workflows"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UpdateWorkflow" } }
          }
        },
        "responses": {
          "200": { "description": "Workflow updated" },
          "404": { "description": "Not found" }
        }
      },
      "delete": {
        "summary": "Delete a workflow",
        "tags": ["Workflows"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Workflow deleted" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/workflows/{id}/run": {
      "post": {
        "summary": "Run a workflow",
        "tags": ["Workflows"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "202": { "description": "Workflow run started" },
          "400": { "description": "Cannot run workflow" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/sessions": {
      "post": {
        "summary": "Create a browser session",
        "tags": ["Sessions"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "browserType": { "type": "string" } } }
            }
          }
        },
        "responses": { "201": { "description": "Session created" } }
      },
      "get": {
        "summary": "List sessions",
        "tags": ["Sessions"],
        "parameters": [{ "name": "status", "in": "query", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "List of sessions" } }
      }
    },
    "/api/sessions/{id}": {
      "get": {
        "summary": "Get session",
        "tags": ["Sessions"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Session details" },
          "404": { "description": "Not found" }
        }
      },
      "delete": {
        "summary": "Close session",
        "tags": ["Sessions"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Session closed" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/credentials": {
      "post": {
        "summary": "Store credentials",
        "tags": ["Credentials"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateCredential" } }
          }
        },
        "responses": {
          "201": { "description": "Credential stored" },
          "400": { "description": "Validation error" }
        }
      },
      "get": {
        "summary": "List credentials",
        "tags": ["Credentials"],
        "responses": { "200": { "description": "List of credentials" } }
      }
    },
    "/api/dashboard": {
      "get": {
        "summary": "Dashboard snapshot",
        "tags": ["Dashboard"],
        "responses": { "200": { "description": "Full state snapshot" } }
      }
    },
    "/api/dashboard/stream": {
      "get": {
        "summary": "SSE event stream",
        "tags": ["Dashboard"],
        "responses": { "200": { "description": "Server-sent events stream" } }
      }
    },
    "/api/dashboard/run": {
      "post": {
        "summary": "Spawn test run(s)",
        "tags": ["Dashboard"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/SpawnRun" } }
          }
        },
        "responses": {
          "201": { "description": "Runs spawned" },
          "400": { "description": "Validation error" }
        }
      }
    },
    "/api/devices": {
      "get": {
        "summary": "Device presets",
        "tags": ["System"],
        "responses": { "200": { "description": "List of device presets" } }
      }
    },
    "/api/agents": {
      "get": {
        "summary": "AI agents",
        "tags": ["System"],
        "responses": { "200": { "description": "List of available agents" } }
      }
    },
    "/api/metrics": {
      "get": {
        "summary": "Prometheus metrics",
        "tags": ["System"],
        "security": [],
        "responses": { "200": { "description": "Prometheus text format metrics" } }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "CreateTask": {
        "type": "object",
        "required": ["prompt", "url"],
        "properties": {
          "prompt": { "type": "string", "minLength": 1 },
          "url": { "type": "string", "format": "uri" },
          "maxSteps": { "type": "integer", "minimum": 1, "maximum": 100, "default": 25 },
          "maxIterations": { "type": "integer", "minimum": 1, "maximum": 50, "default": 10 },
          "webhookCallbackUrl": { "type": "string", "format": "uri" },
          "errorCodes": { "type": "object", "additionalProperties": { "type": "string" } },
          "extractionSchema": { "type": "object" },
          "navigationPayload": { "type": "object" },
          "totpCredentialId": { "type": "string" }
        }
      },
      "TaskResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "status": {
            "type": "string",
            "enum": ["queued", "running", "completed", "failed", "cancelled"]
          },
          "createdAt": { "type": "number" }
        }
      },
      "CreateWorkflow": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "minLength": 1 },
          "description": { "type": "string" },
          "status": {
            "type": "string",
            "enum": ["draft", "active", "paused", "archived"],
            "default": "draft"
          },
          "blocks": { "type": "array", "default": [] },
          "parameters": { "type": "object" },
          "cronSchedule": { "type": "string" },
          "strictMode": { "type": "boolean", "default": false },
          "tags": { "type": "array", "items": { "type": "string" } }
        }
      },
      "UpdateWorkflow": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1 },
          "description": { "type": "string" },
          "status": { "type": "string", "enum": ["draft", "active", "paused", "archived"] },
          "blocks": { "type": "array" },
          "parameters": { "type": "object" },
          "cronSchedule": { "type": "string" },
          "strictMode": { "type": "boolean" },
          "tags": { "type": "array", "items": { "type": "string" } }
        }
      },
      "CreateCredential": {
        "type": "object",
        "required": ["type", "label"],
        "properties": {
          "type": {
            "type": "string",
            "enum": ["password", "api-key", "oauth", "totp", "certificate"]
          },
          "label": { "type": "string", "minLength": 1 },
          "provider": {
            "type": "string",
            "enum": ["native", "bitwarden", "1password", "azure-key-vault", "custom-http"],
            "default": "native"
          },
          "data": { "type": "object" },
          "domain": { "type": "string" }
        }
      },
      "SpawnRun": {
        "type": "object",
        "required": ["instruction", "devices"],
        "properties": {
          "instruction": { "type": "string", "minLength": 1 },
          "url": { "type": "string", "format": "uri" },
          "agent": { "type": "string" },
          "mode": { "type": "string", "enum": ["dom", "hybrid", "cua"] },
          "browser": { "type": "string", "enum": ["chromium", "firefox", "webkit"] },
          "devices": { "type": "array", "items": { "type": "string" }, "minItems": 1 },
          "headed": { "type": "boolean", "default": false },
          "a11y": { "type": "boolean", "default": false },
          "lighthouse": { "type": "boolean", "default": false }
        }
      },
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["healthy", "degraded", "unhealthy"] },
          "uptime": { "type": "number" },
          "checks": { "type": "array", "items": { "type": "object" } }
        }
      }
    }
  }
}